requirements.txt
Django==2.2.6
channels==2.3.0
channels-redis==2.4.0
django-cors-headers==3.1.1
django-extensions==2.2.3
django-filter==2.2.0
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0